+++ lib/api/registry.dart
@@ -7,34 +7,27 @@ import '../models/dart/entity.pb.dart';
 Future<Map<String, dynamic>> validateRegistrationToken(
     EntityReference entityRef,
     String token,
-    Map<String, dynamic> userData,
+    Map<String, dynamic> userInfo,
     DVoteGateway dvoteGw,
     String privateKey) async {
-  if (!(userData is Map) || !(dvoteGw is DVoteGateway) || (dvoteGw == null)) {
+  if (!(userInfo is Map) || !(dvoteGw is DVoteGateway) || (dvoteGw == null)) {
     throw Exception("Invalid parameters");
   } else if (!(token is String) ||
-      !(userData["firstName"] is String) ||
-      !(userData["lastName"] is String) ||
-      !(userData["email"] is String) ||
-      !(userData["phone"] is String) ||
-      !(userData["dateOfBirth"] is String) ||
+      !(userInfo["firstName"] is String) ||
+      !(userInfo["lastName"] is String) ||
+      !(userInfo["email"] is String) ||
+      !(userInfo["phone"] is String) ||
+      !(userInfo["dateOfBirth"] is String) ||
       !(privateKey is String && privateKey != "")) {
     throw Exception("Invalid parameters");
   }
 
   try {
-    Map<String, dynamic> memberInfo = {
-      "firstName": userData["firstName"],
-      "lastName": userData["lastName"],
-      "email": userData["email"],
-      "phone": userData["phone"],
-      "dateOfBirth": userData["dateOfBirth"],
-    };
     Map<String, dynamic> reqParams = {
       "method": "register",
       "token": token,
       "entityId": entityRef.entityId,
-      "member": memberInfo
+      "member": userInfo
     };
     Map<String, dynamic> response = await dvoteGw.sendRequest(reqParams,
         timeout: 7, privateKey: privateKey);