+++ lib/api/registry.dart
@@ -1,33 +1,54 @@
 import 'package:dvote/dvote.dart';
-
 import '../net/gateway.dart';
-import '../models/dart/entity.pb.dart';
 
 // HANDLERS
-Future<Map<String, dynamic>> validateRegistrationToken(
-    EntityReference entityRef,
-    String token,
-    Map<String, dynamic> userInfo,
+Future<Map<String, dynamic>> register(
+    String entityId,
+    String firstName,
+    String lastName,
+    String email,
+    String phone,
+    DateTime dateOfBirth,
     DVoteGateway dvoteGw,
     String privateKey) async {
-  if (!(userInfo is Map) || !(dvoteGw is DVoteGateway) || (dvoteGw == null)) {
-    throw Exception("Invalid parameters");
-  } else if (!(token is String) ||
-      !(userInfo["firstName"] is String) ||
-      !(userInfo["lastName"] is String) ||
-      !(userInfo["email"] is String) ||
-      !(userInfo["phone"] is String) ||
-      !(userInfo["dateOfBirth"] is String) ||
-      !(privateKey is String && privateKey != "")) {
+  try {
+    Map<String, dynamic> reqParams = {
+      "method": "register",
+      "member": {
+        "firstName": firstName,
+        "lastName": lastName,
+        "email": email,
+        "phone": phone,
+        "dateOfBirth": dateOfBirth.toUtc().toIso8601String(),
+      },
+      "entityId": entityId
+    };
+    Map<String, dynamic> response = await dvoteGw.sendRequest(reqParams,
+        timeout: 7, privateKey: privateKey);
+    if (!(response is Map)) {
+      throw Exception("Invalid response received from the gateway");
+    }
+    return response;
+  } catch (err) {
+    throw Exception(
+        "The registration token could not be validated: " + err.toString());
+  }
+}
+
+Future<Map<String, dynamic>> validateRegistrationToken(String entityId,
+    String validationToken, DVoteGateway dvoteGw, String privateKey) async {
+  if (!(entityId is String) ||
+      !(validationToken is String) ||
+      !(dvoteGw is DVoteGateway) ||
+      !(privateKey is String)) {
     throw Exception("Invalid parameters");
   }
 
   try {
     Map<String, dynamic> reqParams = {
-      "method": "register",
-      "token": token,
-      "entityId": entityRef.entityId,
-      "member": userInfo
+      "method": "validateToken",
+      "token": validationToken,
+      "entityId": entityId
     };
     Map<String, dynamic> response = await dvoteGw.sendRequest(reqParams,
         timeout: 7, privateKey: privateKey);